+++ tests/provision/bootc/test.sh
@@ -31,7 +31,7 @@ rlJournalStart
         rlRun -s "tmt -vvv run --scratch -i $run plan --name /plans/image/needs-deps"
         # Testing the output of the bootc package manager
         rlAssertGrep "building container image with dependencies" $rlRun_LOG
-        rlAssertGrep "STEP 1/2: FROM containers-storage:localhost/tmt/bootc" $rlRun_LOG
+        rlAssertGrep "STEP 1/2: FROM containers-storage:localhost/tmtmodified" $rlRun_LOG
         rlAssertGrep "Successfully tagged localhost/tmt/bootc/" $rlRun_LOG
         rlAssertGrep "switching to new image localhost/tmt/bootc/" $rlRun_LOG
         rlAssertGrep "rebooting to apply new image" $rlRun_LOG