+++ tmt/steps/__init__.py
@@ -2,12 +2,14 @@
 Step Classes
 """
 
+import abc
 import collections
 import functools
 import itertools
 import re
 import shutil
 import textwrap
+import typing
 from collections.abc import Iterable, Iterator, Sequence
 from contextlib import suppress
 from re import Pattern
@@ -16,6 +18,7 @@ from typing import (
     Any,
     Callable,
     Generic,
+    Literal,
     Optional,
     TypedDict,
     TypeVar,
@@ -37,7 +40,7 @@ import tmt.options
 import tmt.queue
 import tmt.utils
 import tmt.utils.rest
-from tmt._compat.typing import Self
+from tmt._compat.typing import Self, TypeGuard
 from tmt.container import (
     SerializableContainer,
     SpecBasedContainer,
@@ -103,8 +106,10 @@ PHASE_ORDER_PREPARE_INSTALL_REQUIRES = 70
 PHASE_ORDER_PREPARE_INSTALL_RECOMMENDS = 75
 
 # Supported steps and actions
-STEPS: list[str] = ['discover', 'provision', 'prepare', 'execute', 'report', 'finish', 'cleanup']
-ACTIONS: list[str] = ['login', 'reboot']
+StepName = Literal['discover', 'provision', 'prepare', 'execute', 'report', 'finish', 'cleanup']
+STEPS: list[StepName] = list(typing.get_args(StepName))
+ActionName = Literal['login', 'reboot']
+ACTIONS: list[ActionName] = list(typing.get_args(ActionName))
 DEFAULT_LOGIN_COMMAND = 'bash'
 
 #: A default command to trigger a guest reboot when executed remotely.
@@ -166,6 +171,14 @@ PHASE_OPTIONS = tmt.options.create_options_decorator(
         ),
     ]
 )
+
+
+def is_step_name(val: str) -> TypeGuard[StepName]:
+    return val in STEPS
+
+
+def is_action_name(val: str) -> TypeGuard[ActionName]:
+    return val in ACTIONS
 
 
 def prune_directory(path: Path, preserved_members: set[str], logger: tmt.log.Logger) -> set[str]:
@@ -806,6 +819,7 @@ class Step(
         for data in self.data:
             self._plugin_base_class.delegate(self, data=data).show()
 
+    @abc.abstractmethod
     def summary(self) -> None:
         """
         Give a concise summary about the step result
@@ -1300,7 +1314,7 @@ class Step(
         Run all loaded Login or Reboot action instances of the step
         """
 
-        for phase in self.phases(classes=Action):
+        for phase in self.phases(classes=Action):  # type: ignore[type-abstract]
             phase.go()
 
     def go(self, force: bool = False) -> None:
@@ -1335,7 +1349,7 @@ class Step(
 
         # Do not prune plugin workdirs, each plugin decides what should
         # be pruned from the workdir and what should be kept there
-        plugins = self.phases(classes=BasePlugin)
+        plugins = self.phases(classes=BasePlugin)  # type: ignore[type-abstract]
         for plugin in plugins:
             if plugin.workdir is not None:
                 preserved_members = {*preserved_members, plugin.workdir.name}
@@ -1602,6 +1616,7 @@ class BasePlugin(
         return self.pathless_safe_name
 
     @classmethod
+    @abc.abstractmethod
     def base_command(
         cls,
         usage: str,
@@ -2067,6 +2082,7 @@ class GuestlessPlugin(BasePlugin[StepDataT, PluginReturnValueT]):
     Common parent of all step plugins that do not work against a particular guest
     """
 
+    @abc.abstractmethod
     def go(self, *, logger: Optional[tmt.log.Logger] = None) -> PluginReturnValueT:
         """
         Perform actions shared among plugins when beginning their tasks
@@ -2080,6 +2096,7 @@ class Plugin(BasePlugin[StepDataT, PluginReturnValueT]):
     Common parent of all step plugins that do work against a particular guest
     """
 
+    @abc.abstractmethod
     def go(
         self,
         *,
@@ -2171,6 +2188,7 @@ class Action(Phase, tmt.utils.MultiInvokableCommon):
                 phases[step_name] = [phase]
         return phases
 
+    @abc.abstractmethod
     def go(self) -> None:
         raise NotImplementedError
 
@@ -2426,7 +2444,7 @@ class Login(Action):
         for guest in self.parent.plan.provision.ready_guests:
             # Attempt to push the workdir to the guest
             try:
-                guest.push()
+                guest.push(source=self.parent.plan.worktree)
                 if not cwd:
                     # Use path of the last executed test as the default
                     # current working directory
@@ -2437,6 +2455,7 @@ class Login(Action):
                         cwd = worktree
                     else:
                         try:
+                            guest.push(source=self.parent.plan.discover.workdir)
                             cwd = worktree.parent / "discover" / test_path.unrooted()
                             guest.execute(
                                 tmt.utils.ShellScript("/bin/true"),