+++ tmt/steps/provision/testcloud.py
@@ -121,8 +121,10 @@ chpasswd:
 users:
   - default
   - name: ${user_name}
-ssh_authorized_keys:
-  - ${public_key}
+    groups: [wheel, sudo]
+    sudo: ['ALL=(ALL) NOPASSWD:ALL']
+    ssh_authorized_keys:
+      - ${public_key}
 ssh_pwauth: true
 disable_root: false
 runcmd:
@@ -134,6 +136,7 @@ version: 1.4.0
 passwd:
   users:
     - name: ${user_name}
+      groups: [wheel, sudo]
       ssh_authorized_keys:
         - ${public_key}
 storage:
@@ -146,6 +149,12 @@ storage:
           # Enable it.
           # This file must sort before 40-rhcos-defaults.conf.
           PermitRootLogin yes
+    - path: /etc/sudoers.d/tmt-user
+      mode: 0440
+      contents:
+        inline: |
+          # Allow tmt user to run sudo commands without password
+          ${user_name} ALL=(ALL) NOPASSWD:ALL
 """
 
 # VM defaults
@@ -1008,8 +1017,23 @@ class GuestTestcloud(tmt.GuestSsh):
             self.image_url = self._guess_image_url(self.image_url)
             self.debug(f"Guessed image url: '{self.image_url}'", level=3)
 
+        assert testcloud is not None  # Narrow type
+
+        # Make a symlink to the file image
+        if self.image_url.startswith("file://"):
+            image_path = Path(self.image_url.removeprefix("file://"))
+            # We should not symlink any supported formats, e.g. the `.xz`
+            # does an extract step that would be skip if we make the symlink.
+            if image_path.suffixes and image_path.suffixes[-1] in (".qcow2",):
+                # Create a symlink in the testcloud STORE_DIR and make sure
+                # it is always updated to the requested version.
+                image_symlink = self.testcloud_image_dirpath / image_path.name
+                image_symlink.unlink(missing_ok=True)
+                image_symlink.symlink_to(image_path)
+                # Adjust selinux tags for the actual image
+                testcloud.image.Image._adjust_image_selinux(image_path)
+
         # Initialize and prepare testcloud image
-        assert testcloud is not None
         self._image = testcloud.image.Image(self.image_url)
         self.verbose('qcow', self._image.name, 'green')
         if not Path(self._image.local_path).exists():