+++ docs/releases.rst
@@ -4,6 +4,14 @@
     Releases
 ======================
 
+tmt-1.61.0
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Polarion export plugin now fails properly when exporting an fmf tree that
+is not part of a git repository and when there are unsubmitted changes.
+Also allowing the export when ``--ignore-git-validation`` is passed.
+
+
 tmt-1.60.0
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -15,9 +23,61 @@ While running ``tmt plan export``, the exported plan now includes
 combined context from multiple sources, including command line,
 making it consistent with ``tmt plan show`` command.
 
-Polarion export plugin now fails properly when exporting an fmf tree that
-is not part of a git repository and when there are unsubmitted changes.
-Also allowing the export when ``--ignore-git-validation`` is passed.
+New :ref:`sprints` section has been added to the documentation for
+contributors. It describes the overall sprint process, the key
+sync points and provides links to progress tracking boards.
+
+tmt now supports automatic generation of Ansible inventory files
+during the provision phase. The new :ref:`/spec/plans/ansible`
+configuration allows users to define custom inventory layouts and
+organize provisioned guests into Ansible groups with host-specific
+variables. See the :ref:`plan ansible </spec/plans/ansible>` and
+:ref:`provision ansible </spec/plans/provision/ansible>` keys for
+configuration details.
+
+The :ref:`/plugins/report/reportportal` report plugin now supports
+a new ``auto-analysis`` key. This key allows users to enable
+immediate auto-analysis of failed tests reported to ReportPortal.
+
+User generated content in the ``TMT_PLAN_DATA`` directory is now
+correctly preserved when the :ref:`/stories/cli/run/login` action
+is used to log into the guest.
+
+The :ref:`system.type</spec/hardware/system>` HW requirement is now
+recognized by tmt.
+
+Beakerlib output from phases is now saved in files with more predictable
+filenames, ``output.txt`` instead of random temporary filenames.
+
+The ``bootc`` package manager now correctly checks for the existence of
+installable packages, and skips their installation - and image rebuild
+- when they are installed already.
+
+``tmt cleanup`` now correctly handles workdirs that no longer contain a
+valid ``run.yaml`` file. This allows guest removal in the case of invalid
+or incomplete runs.
+
+The :ref:`/plugins/provision/bootc` guest now uses ``containers-storage:``
+prefix for local containers only. It was incorrectly used for remote
+containers, and the image build process might have failed when such an
+image was not already present on the guest.
+
+SSH key placement and non-root ``sudo`` have been fixed in the
+:ref:`/plugins/provision/virtual.testcloud` guests.
+
+:ref:`/spec/policy` examples were updated to demonstrate the use of
+``to_yaml`` filter with complex objects, which preserves them correctly.
+
+The implementation of the :ref:`boot.method</spec/hardware/boot>` HW
+requirement changes to support instance types which offer multiple boot
+methods at the same time. Plugin is then free to choose the right one,
+given the HW requirements. Current strict implementation would prevent
+the use of such instance types.
+
+:ref:`/plugins/provision` which use ``ssh`` to connect to the
+guest, for example :ref:`/plugins/provision/connect` or
+:ref:`/plugins/provision/virtual.testcloud`, now support providing
+multiple ssh keys using the ``--key`` option.
 
 
 tmt-1.59.0